Extract leaderboard enums
