+++ src/enums/option_types/leaderboard_duration.ts
@@ -1,0 +1,11 @@
+export enum LeaderboardDuration {
+    TODAY = "today",
+    DAILY = "daily",
+    WEEK = "week",
+    WEEKLY = "weekly",
+    MONTH = "month",
+    MONTHLY = "monthly",
+    YEAR = "year",
+    YEARLY = "yearly",
+    ALL_TIME = "alltime",
+}
