+++ src/enums/option_types/leaderboard_scope.ts
@@ -1,0 +1,5 @@
+export enum LeaderboardScope {
+    GLOBAL = "global",
+    SERVER = "server",
+    GAME = "game",
+}
