+++ src/enums/option_types/leaderboard_type.ts
@@ -1,0 +1,5 @@
+export enum LeaderboardType {
+    EXP = "exp",
+    GAMES_PLAYED = "gamesplayed",
+    SONGS_GUESSED = "songsguessed",
+}
